main,appengine: split main package into standalone and appengine
