+++ main/main.go
@@ -1,5 +1,3 @@
-// +build !appengine
-
 package main
 
 import (
@@ -10,6 +8,7 @@ import (
 	"github.com/shirasudon/go-chat/domain"
 	"github.com/shirasudon/go-chat/infra/inmemory"
 	"github.com/shirasudon/go-chat/infra/pubsub"
+	"github.com/shirasudon/go-chat/main/resolve"
 )
 
 type DoneFunc func()
@@ -45,7 +44,7 @@ func createInfra() (domain.Repositories, *chat.Queryers, chat.Pubsub, DoneFunc) 
 
 func main() {
 	repos, qs, ps, infraDoneFunc := createInfra()
-	s, done := createServer(repos, qs, ps)
+	s, done := resolve.CreateServer(repos, qs, ps)
 	defer func() {
 		done()
 		infraDoneFunc()
