+++ main/resolve/resolve.go
@@ -1,15 +1,26 @@
-package main
+// package resolve provide helper functions for resolving dependencies to
+// construct gochat Server.
+
+package resolve
 
 import (
 	"context"
+	"fmt"
 
 	gochat "github.com/shirasudon/go-chat"
 	"github.com/shirasudon/go-chat/chat"
 	"github.com/shirasudon/go-chat/domain"
 )
 
-// createServer creates server with resolve its dependencies.
-func createServer(repos domain.Repositories, qs *chat.Queryers, ps chat.Pubsub) (server *gochat.Server, done DoneFunc) {
+func main() {
+	fmt.Println("vim-go")
+}
+
+// DoneFunc is function to be called after all of operations are done.
+type DoneFunc func()
+
+// CreateServer creates server with resolve its dependencies.
+func CreateServer(repos domain.Repositories, qs *chat.Queryers, ps chat.Pubsub) (server *gochat.Server, done DoneFunc) {
 	chatCmd := chat.NewCommandServiceImpl(repos, ps)
 	chatQuery := chat.NewQueryServiceImpl(qs)
 	chatHub := chat.NewHubImpl(chatCmd)
